+++ packages/frameworks/solid/src/file-upload/file-upload.test.tsx
@@ -42,6 +42,7 @@ const ComponentUnderTest = (props: FileUploadProps) => (
 describe('FileUpload', () => {
   it.each(getParts(fileUploadAnatomy))('should render part! %s', async (part) => {
     render(() => <ComponentUnderTest />)
+    // eslint-disable-next-line testing-library/no-node-access
     expect(document.querySelector(part)).toBeInTheDocument()
   })
 })