+++ packages/frameworks/solid/src/hover-card/hover-card.test.tsx
@@ -25,6 +25,7 @@ const ComponentUnderTest = (props: HoverCardProps) => (
 describe('HoverCard', () => {
   it.each(getParts(hoverCardAnatomy))('should render part! %s', async (part) => {
     render(() => <ComponentUnderTest />)
+    // eslint-disable-next-line testing-library/no-node-access
     expect(document.querySelector(part)).toBeInTheDocument()
   })
 
@@ -59,7 +60,7 @@ describe('HoverCard', () => {
     expect(screen.queryByTestId('positioner')).not.toBeInTheDocument()
 
     await user.hover(screen.getByText('Hover me'))
-    expect(screen.queryByTestId('positioner')).toBeInTheDocument()
+    expect(screen.getByTestId('positioner')).toBeInTheDocument()
   })
 
   it('should lazy mount and unmount on exit', async () => {
@@ -67,7 +68,7 @@ describe('HoverCard', () => {
     expect(screen.queryByTestId('positioner')).not.toBeInTheDocument()
 
     await user.hover(screen.getByText('Hover me'))
-    expect(screen.queryByTestId('positioner')).toBeInTheDocument()
+    expect(screen.getByTestId('positioner')).toBeInTheDocument()
 
     await user.unhover(screen.getByText('Hover me'))
     await waitFor(() => expect(screen.queryByTestId('positioner')).not.toBeInTheDocument())