+++ packages/frameworks/solid/src/radio-group/radio-group.test.tsx
@@ -32,6 +32,7 @@ const ComponentUnderTest = (props: RadioGroupProps) => {
 describe('Radio Group', () => {
   it.each(getParts(radioGroupAnatomy))('should render part! %s', async (part) => {
     render(() => <ComponentUnderTest />)
+    // eslint-disable-next-line testing-library/no-node-access
     expect(document.querySelector(part)).toBeInTheDocument()
   })