+++ packages/frameworks/vue/src/accordion/tests/accordion.test.tsx
@@ -7,6 +7,7 @@ import ComponentUnderTest from './accordion.test.vue'
 describe('Accordion', () => {
   it.each(getParts(accordionAnatomy))('should render part %s', async (part) => {
     render(ComponentUnderTest)
+    // eslint-disable-next-line testing-library/no-node-access
     expect(document.querySelector(part)).toBeInTheDocument()
   })
 
@@ -50,7 +51,7 @@ describe('Accordion', () => {
 
   it.skip('should disable all items when disabled is true', async () => {
     render(ComponentUnderTest, { props: { disabled: true } })
-    expect(screen.getByRole('button', { name: 'React Trigger' })).toHaveAttribute('disabled')
+    expect(screen.getByRole('button', { name: 'React Trigger' })).toBeDisabled()
   })
 
   it('should allow multiple items to be expanded when multiple is true', async () => {