+++ packages/frameworks/react/CHANGELOG.md
@@ -6,9 +6,18 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
+## [1.2.1] - 2023-12-13
+
+### Fixed
+
+- Resolved an issue where the `ark` factory type was not being exported correctly.
+
+## [1.2.0] - 2023-12-13
+
 ### Added
 
 - Added the `ToastGroup` component.
+- Added entrypoint for the  `ark` factory at `@ark-ui/react/factory`
 
 ### Changed
 
@@ -20,6 +29,7 @@ description: All notable changes to this project will be documented in this file
 - Resolved an issue where the `present` prop in the disclosure-type component was not being respected.
 - Resolved an issue where the `ark` function would log a warning when the `asChild` prop was set to `false`.
 - Fixed an issue where keyboard interactions within a submenu would bubble up to the parent `Menu`.
+- Fixed an issue with hydration mismatch in the `Portal` component.
 
 ## [1.1.0] - 2023-11-21